backend/requirements.txt
# ── Web framework ─────────────────────────────────────────────
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12     # form data support

# ── Database ──────────────────────────────────────────────────
supabase==2.10.0
pgvector==0.3.0

# ── Auth / Security ───────────────────────────────────────────
python-jose[cryptography]==3.3.0
bcrypt>=4.0.0

# ── Config ────────────────────────────────────────────────────
pydantic-settings>=2.6.0
pydantic[email]>=2.10.0

# ── HTTP client ───────────────────────────────────────────────
httpx==0.27.0

# ── AI (Sprint 2+) ────────────────────────────────────────────
openai>=1.50.0              # embeddings (text-embedding-3-small)
anthropic>=0.40.0           # Claude for KB parsing
firecrawl-py>=1.0.0         # website scraping

# ── AI (added Sprint 3+) ──────────────────────────────────────
# langchain==0.3.0
# langchain-anthropic==0.3.0
# langchain-openai==0.2.0
# langchain-community==0.3.0
# crewai==0.80.0
# crewai-tools==0.14.0

# ── Messaging (added Sprint 7+) ───────────────────────────────
# twilio==9.3.0
# resend==2.5.0

# ── Payments (added Sprint 7) ─────────────────────────────────
# stripe==11.1.0

# ── Background jobs (added Sprint 11) ────────────────────────
# celery==5.4.0
# redis==5.2.0
